requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-huggingface
langchain-groq

chromadb

sentence-transformers
transformers
torch